Fix small styling changes (#164)
Added 5 pixel to room name input box so logo is no longer mixing with the label.
Updated tranlsations that were not in the language files + created script to look for occurances. (also added some missing hu translations)
Update labels for media device chooser, it is no longer mixing with the shadow of the media device preview.